chore(examples): remove dynamic versions of example apps
