+++ modules/examples/e2e_test/hello_world/hello_world_spec.ts
@@ -4,25 +4,8 @@ describe('hello world', function() {
 
   afterEach(verifyNoBrowserErrors);
 
-  describe('static reflection', function() {
+  describe('hello world app', function() {
     var URL = 'examples/src/hello_world/index.html';
-
-    it('should greet', function() {
-      browser.get(URL);
-
-      expect(getComponentText('hello-app', '.greeting')).toEqual('hello world!');
-    });
-
-    it('should change greeting', function() {
-      browser.get(URL);
-
-      clickComponentButton('hello-app', '.changeButton');
-      expect(getComponentText('hello-app', '.greeting')).toEqual('howdy world!');
-    });
-  });
-
-  describe('dynamic reflection', function() {
-    var URL = 'examples/src/hello_world/index_dynamic.html';
 
     it('should greet', function() {
       browser.get(URL);
