--- modules/examples/src/http/index_dynamic.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html>
-  <title>Angular 2.0 Http (Reflection)</title>
-<body>
-  <http-app>
-    Loading...
-  </http-app>
-
-  $SCRIPTS$
-</body>
-</html>
