--- modules/examples/src/jsonp/index_dynamic.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html>
-  <title>Angular 2.0 Jsonp (Reflection)</title>
-<body>
-  <jsonp-app>
-    Loading...
-  </jsonp-app>
-
-  $SCRIPTS$
-</body>
-</html>
